Add a "no_connector" connector as a proof of concept that different connectors will work.
The no_connector always returns nothing when asking the mis for data but
should still work fine for data already imported or manually input.
Could be a good thing to switch to when mis is down, or to use for
testing.

It's possible that, in the future, forms could be created to populate
the ilp database instead of requiring an MIS