Refactored out some stuff from view#show
Get date now sets @date in a before filter. This means I don't have to specifically do it in view#show and in the future I'll make timetables use it too.
Dropped the affiliation scope for views and replaced with with for_user which scopes for the current users' affiliation
Simplified a load of stuff in the views controller and view - mostly to do with refactoring into the the two things above and simplifying how the future/past line is worked out.
Fixed a bug where the everyone sidebar could show for students (although they'd never be able to see any actual events)