Leap 5.2.1
Rails version - 5.2.1
Ruby version - 2.3.3

This is largely in a working state now. Styling and layout seem to be correct. I've done a lot of testing and worked through a number of syntax issues, though it's likely more will crop up.
On the older style pages the admin menu dropdown isn't working for some reason.